pyproject.toml
[project]
name = "zeno-fastapi-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomysql>=0.2.0",
    "asyncpg>=0.30.0",
    "bcrypt==3.2.2",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
